login page + login controller
The login page displays error messages in case of an invalid format of the user input. The login controller contains a function that directs to the login form and another one that checks if the user input has the correct format. In the case that it does, it directs the user to a loginTest page. Otherwise, it redirects them to the login form, which now contains error messages.